requirements.txt
tensordict-nightly
torchrl-nightly
pymap3d==3.0.1
geographiclib==2.0
gymnasium==1.0.0a1
wandb==0.15.12
icecream==2.1.3
shapely==2.0.2
jsbsim==1.2.0
matplotlib==3.8.1
hydra-core==1.3.2
tqdm==4.66.1
glfw==2.7.0
mujoco==3.1.4